feat: Add TrainerStatus collector and UI progress tracking
- Add TrainerStatus metrics collector for TrainJob progress monitoring
- Fix EstimatedRemainingSeconds type to int64 (matches TrainJob CRD)
- Add TrialsProgress and ConvergenceChart UI components
- Update experiment-details page with real-time progress visualization
- Add TrainerStatus option in metrics collector UI selection
- Update E2E tests for TrainerStatus collector
- Add local development script

+++ pkg/apis/controller/common/v1beta1/common_types.go
@@ -169,8 +169,8 @@ type TrainingProgress struct {
 	// Current progress percentage (0-100)
 	ProgressPercentage int32 `json:"progressPercentage,omitempty"`
 
-	// Estimated remaining time in seconds
-	EstimatedRemainingSeconds int32 `json:"estimatedRemainingSeconds,omitempty"`
+	// Estimated remaining time in seconds (matches TrainJob CRD int64 type)
+	EstimatedRemainingSeconds int64 `json:"estimatedRemainingSeconds,omitempty"`
 
 	// Current training step/iteration
 	CurrentStep int32 `json:"currentStep,omitempty"`
